requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
Pygments==2.18.0
sqlparse==0.5.0
pytest==8.3.3
pytest-django==4.9.0
factory_boy==3.3.1
pytest-factoryboy==2.7.0